requirements.txt
click==7.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.3
numpy==1.18.2
opencv-contrib-python==4.2.0.32
Pillow==7.0.0
scikit-learn==0.22.2.post1
scipy==1.4.1
sklearn==0.0
pandas==1.0.3
pylab-sdk==1.1.2
requests==2.23.0